Extract deprecated checkBool
